Moved the container model out of models.
This is as its not currently a proper model.
