+++ provisions/roles/jenkins/common/tasks/main.yml
@@ -8,15 +8,15 @@
   - jenkins/slave
   - config
 
-- name: Ensure pip is installed.
-  yum: name=python-pip state=present
-  tags:
-    - jenkins
-    - jenkins/master
-    - jenkins/slave
+    #- name: Ensure pip is installed.
+    #  yum: name=python-pip state=present
+    #  tags:
+    #    - jenkins
+    #    - jenkins/master
+    #    - jenkins/slave
 
-- name: Ensure django is installed
-  shell: "pip install django"
+      #- name: Ensure django is installed
+      #  shell: "pip install django"
 
 - name: ensure java-1.7.0-* is not installed
   yum: name=java-1.7.0-* state=absent
